requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
apispec==6.6.1
attrs==24.2.0
bech32==1.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cosmpy==0.9.2
customtkinter==5.2.2
darkdetect==0.8.0
ecdsa==0.19.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.3
langchain-core==0.3.10
langchain-google-genai==2.0.1
langchain-text-splitters==0.3.0
langsmith==0.1.134
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pillow==10.4.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
uagents==0.16.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
websockets==10.4
yarl==1.14.0
